Extract stable extension seams for extensions
