+++ .github/workflows/build-and-test.yml
@@ -33,6 +33,9 @@ jobs:
       - name: Grant execute permission for Gradlew
         run: chmod +x gradlew
 
+      - name: Verify Extension Seams
+        run: bash scripts/ci/check-extension-seams.sh
+
       - name: Build and Test with Coverage
         run: ./gradlew clean build koverXmlReport --stacktrace
