+++ core/src/commonTest/kotlin/contract/ReadCoreContractTest.kt
@@ -6,18 +6,18 @@ import dev.mattramotar.storex.core.Freshness
 import dev.mattramotar.storex.core.StoreKey
 import dev.mattramotar.storex.core.StoreResult
 import dev.mattramotar.storex.core.TimeSource
-import dev.mattramotar.storex.core.internal.Bookkeeper
-import dev.mattramotar.storex.core.internal.ConditionalRequest
-import dev.mattramotar.storex.core.internal.DefaultDbMeta
-import dev.mattramotar.storex.core.internal.DefaultFreshnessValidator
-import dev.mattramotar.storex.core.internal.Fetcher
-import dev.mattramotar.storex.core.internal.FetchPlan
-import dev.mattramotar.storex.core.internal.FreshnessContext
-import dev.mattramotar.storex.core.internal.FreshnessValidator
-import dev.mattramotar.storex.core.internal.MemoryCache
+import dev.mattramotar.storex.core.seams.Bookkeeper
+import dev.mattramotar.storex.core.seams.ConditionalRequest
+import dev.mattramotar.storex.core.seams.DefaultDbMeta
+import dev.mattramotar.storex.core.seams.DefaultFreshnessValidator
+import dev.mattramotar.storex.core.seams.Fetcher
+import dev.mattramotar.storex.core.seams.FetchPlan
+import dev.mattramotar.storex.core.seams.FreshnessContext
+import dev.mattramotar.storex.core.seams.FreshnessValidator
+import dev.mattramotar.storex.core.seams.MemoryCache
 import dev.mattramotar.storex.core.internal.MemoryCacheImpl
 import dev.mattramotar.storex.core.internal.RealReadStore
-import dev.mattramotar.storex.core.internal.SourceOfTruth
+import dev.mattramotar.storex.core.seams.SourceOfTruth
 import dev.mattramotar.storex.core.utils.ARTICLE_KEY_1
 import dev.mattramotar.storex.core.utils.FakeBookkeeper
 import dev.mattramotar.storex.core.utils.FakeFetcher
