+++ mutations/src/commonMain/kotlin/dev/mattramotar/storex/mutations/internal/RealMutationStore.kt
@@ -10,15 +10,15 @@ import dev.mattramotar.storex.core.SingleFlight
 import dev.mattramotar.storex.core.StoreKey
 import dev.mattramotar.storex.core.StoreNamespace
 import dev.mattramotar.storex.core.StoreResult
-import dev.mattramotar.storex.core.internal.Bookkeeper
-import dev.mattramotar.storex.core.internal.FetchPlan
-import dev.mattramotar.storex.core.internal.FetchRequest
-import dev.mattramotar.storex.core.internal.Fetcher
-import dev.mattramotar.storex.core.internal.FetcherResult
-import dev.mattramotar.storex.core.internal.FreshnessContext
-import dev.mattramotar.storex.core.internal.FreshnessValidator
-import dev.mattramotar.storex.core.internal.MemoryCache
-import dev.mattramotar.storex.core.internal.SourceOfTruth
+import dev.mattramotar.storex.core.seams.Bookkeeper
+import dev.mattramotar.storex.core.seams.FetchPlan
+import dev.mattramotar.storex.core.seams.FetchRequest
+import dev.mattramotar.storex.core.seams.Fetcher
+import dev.mattramotar.storex.core.seams.FetcherResult
+import dev.mattramotar.storex.core.seams.FreshnessContext
+import dev.mattramotar.storex.core.seams.FreshnessValidator
+import dev.mattramotar.storex.core.seams.MemoryCache
+import dev.mattramotar.storex.core.seams.SourceOfTruth
 import dev.mattramotar.storex.mutations.CreatePolicy
 import dev.mattramotar.storex.mutations.CreateResult
 import dev.mattramotar.storex.mutations.DeleteClient
