+++ normalization/runtime/src/commonMain/kotlin/dev/mattramotar/storex/normalization/internal/builder.kt
@@ -1,19 +1,17 @@
 package dev.mattramotar.storex.normalization.internal
 
-// TODO: This file will be completed in Phase 3 (Mutations Module Migration)
-// It requires RealStore and Updater which will be migrated from :store to :mutations
+// TODO: This file will be completed in Phase 4 once the mutation/runtime seams are fully
+// available for normalized builders without depending on legacy concrete runtime types.
 
 /*
 import dev.mattramotar.storex.normalization.schema.SchemaRegistry
 import dev.mattramotar.storex.core.Converter
 import dev.mattramotar.storex.core.StoreKey
-import dev.mattramotar.storex.core.internal.Bookkeeper
-import dev.mattramotar.storex.core.internal.Fetcher
-import dev.mattramotar.storex.core.internal.FreshnessValidator
-import dev.mattramotar.storex.core.internal.MemoryCache
-import dev.mattramotar.storex.core.internal.RealStore
-import dev.mattramotar.storex.core.internal.SourceOfTruth
-import dev.mattramotar.storex.core.internal.Updater
+import dev.mattramotar.storex.core.seams.Bookkeeper
+import dev.mattramotar.storex.core.seams.Fetcher
+import dev.mattramotar.storex.core.seams.FreshnessValidator
+import dev.mattramotar.storex.core.seams.MemoryCache
+import dev.mattramotar.storex.core.seams.SourceOfTruth
 import dev.mattramotar.storex.mutations.Creator
 import dev.mattramotar.storex.mutations.MutationStore
 import dev.mattramotar.storex.normalization.GraphProjection
@@ -33,7 +31,7 @@ fun <K : StoreKey, V: Any, Network: Any, Patch, Draft> buildNormalizedEntityStor
     rootResolver: RootResolver<K>,
     fetcher: Fetcher<K, Network>,
     converter: NormalizationConverter<K, V, Network>,
-    updater: Updater<K, Patch, *>? = null,
+    updater: Any? = null,
     creator: Creator<K, Draft, Network>? = null,
     bookkeeper: Bookkeeper<K>,
     validator: FreshnessValidator<K, Any?>,
@@ -43,19 +41,7 @@ fun <K : StoreKey, V: Any, Network: Any, Patch, Draft> buildNormalizedEntityStor
         NormalizedEntitySot(backend, registry, shape, rootResolver)
 
     @Suppress("UNCHECKED_CAST")
-    return RealStore<K, V, GraphProjection<V>, NormalizedWrite<K>, Network, Patch, Draft, Any?, Any?, Any?>(
-        sot = sot,
-        fetcher = fetcher,
-        updater = updater as Updater<K, Patch, Any?>?,
-        creator = creator,
-        deleter = null,
-        putser = null,
-        converter = converter as Converter<K, V, GraphProjection<V>, Network, NormalizedWrite<K>>,
-        encoder = /* your MutationEncoder<Patch, Draft, V, ...> */ error("Provide encoder"),
-        bookkeeper = bookkeeper,
-        validator = validator,
-        memory = memory
-    )
+    return error("Phase 4 normalized mutation builder not yet implemented after seam extraction")
 }
 
 /**
@@ -69,7 +55,7 @@ fun <K : StoreKey, V: Any, Network: Any, Patch, Draft> buildNormalizedListStore(
     itemShape: Shape<V>,
     fetcher: Fetcher<K, Network>,
     converter: NormalizationConverter<K, List<V>, Network>,
-    updater: Updater<K, Patch, *>? = null,
+    updater: Any? = null,
     creator: Creator<K, Draft, Network>? = null,
     bookkeeper: Bookkeeper<K>,
     validator: FreshnessValidator<K, Any?>,
@@ -79,18 +65,6 @@ fun <K : StoreKey, V: Any, Network: Any, Patch, Draft> buildNormalizedListStore(
         NormalizedListSot(backend, index, registry, itemShape)
 
     @Suppress("UNCHECKED_CAST")
-    return RealStore<K, List<V>, GraphProjection<List<V>>, NormalizedWrite<K>, Network, Patch, Draft, Any?, Any?, Any?>(
-        sot = sot,
-        fetcher = fetcher,
-        updater = updater as Updater<K, Patch, Any?>?,
-        creator = creator,
-        deleter = null,
-        putser = null,
-        converter = converter as Converter<K, List<V>, GraphProjection<List<V>>, Network, NormalizedWrite<K>>,
-        encoder = /* your MutationEncoder */ error("Provide encoder"),
-        bookkeeper = bookkeeper,
-        validator = validator,
-        memory = memory
-    )
+    return error("Phase 4 normalized list builder not yet implemented after seam extraction")
 }
 */
